PMT #113025 Search on URL param
This commit changes the database page such that it will search a value
passed in from the URL query string with the param 'q'.
Ex: www.polarhub.org/database/?q=ice

Also note that I clear the url after searching so that I wouldn't have
to add functionality to update the url param with each search. It can
added later if needed.